requirements.txt
pycryptodome==3.14.1
opencv-python==4.5.5.62
numpy==1.22.2
stegano==0.10.2
termcolor==1.1.0
pyfiglet==0.8.post1
streamlit>=1.28.0
scikit-image>=0.20.0
Pillow>=9.0.0
PyWavelets>=1.4.0
scipy>=1.9.0